+++ CHANGELOG.md
@@ -13,4 +13,5 @@ Would like to implement tests and other sustainability features before bumping v
 
 ## 0.1.1
 - Fixed language grammar to have comprehensive highlighting
-- 
+- Added janky HoverProvider for variable types
+- Added split/reverse to hover provider